Created login page and register page
Here we created login.ejs and register.ejs with separate roles for buyers and sellers, allowing users to sign up and log in as either a buyer or a seller. Upon signing up, users have the option to indicate whether they intend to be a seller by checking a checkbox. If the checkbox is checked, the user will have access to features specific to sellers, such as adding products/items for sale. If the checkbox is not checked, the user is defaulted to a buyer role.