starters and Maincourse backend is completed
Now Main Course page and Starters pages are working properly fine.

As a seller, client sign up for an account or log in to add product(s). During product addition, select the checkbox to categorize items as "Main Course" or leave it unchecked for "Starters".

Then user can navigate to the "Starters" or "Main Course" page to view products categorized accordingly.