Updated createProduct.ejs to automatically assign products to Main Course or Starters
 Implement Feature to Automatically Assign Products to Main Course or Starters

This feature enhancement introduces the capability for sellers to easily categorize newly created products/items within a restaurant management system. Specifically, a checkbox option has been implemented to streamline the process of assigning products to either the Main Course or Starters sections.

Upon creation of a new product or item within the system, the seller now has the option to check a checkbox labeled "Add to Main Course." If this checkbox is selected, the product will be automatically categorized and added to the Main Course section. If left unchecked, the product will default to being categorized as a Starter.

This enhancement aims to improve efficiency for sellers by reducing the manual effort required to categorize newly created products. By providing a simple checkbox option, sellers can quickly determine the appropriate categorization for each product, thereby streamlining the overall product management workflow within the system.