Updated Add to Cart Functionality
When a user clicks the Add to Cart button for a product, the product is added to the cart and redirected to the Home page so that user can add more products.
If the user adds the same item multiple times, the system automatically increases the quantity of the product in the cart, instead of adding duplicate entries.
Users can also increase or decrease the quantity of an item directly from the cart page, for this I use axios from node.js.
The cart's total price is updated based on the quantity and price of the products in the cart.